Move functions to utils
